+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/model/PodRevisionTest.java
@@ -8,6 +8,8 @@ import io.fabric8.kubernetes.api.model.ContainerBuilder;
 import io.fabric8.kubernetes.api.model.LabelSelector;
 import io.fabric8.kubernetes.api.model.Pod;
 import io.fabric8.kubernetes.api.model.PodBuilder;
+import io.fabric8.kubernetes.api.model.Quantity;
+import io.fabric8.kubernetes.api.model.ResourceRequirementsBuilder;
 import io.strimzi.api.kafka.model.podset.StrimziPodSet;
 import io.strimzi.api.kafka.model.podset.StrimziPodSetBuilder;
 import io.strimzi.operator.common.Reconciliation;
@@ -33,6 +35,10 @@ public class PodRevisionTest {
                             .withImage("busybox")
                             .withCommand("sleep", "3600")
                             .withImagePullPolicy("IfNotPresent")
+                            .withResources(new ResourceRequirementsBuilder()
+                                    .withRequests(Map.of("cpu", new Quantity("1"), "memory", new Quantity("1024Mi")))
+                                    .withLimits(Map.of("cpu", new Quantity("2"), "memory", new Quantity("2048Mi")))
+                                    .build())
                             .build())
                     .withRestartPolicy("Always")
                     .withTerminationGracePeriodSeconds(0L)
@@ -47,8 +53,21 @@ public class PodRevisionTest {
         assertThat(PodRevision.getRevision(Reconciliation.DUMMY_RECONCILIATION, POD), is(basicPodRevision));
         assertThat(PodRevision.getRevision(Reconciliation.DUMMY_RECONCILIATION, new PodBuilder(POD).build()), is(basicPodRevision));
 
-        // Different pods have different revisions
+        // Revision does not change when resources change
         Pod pod2 = new PodBuilder(POD)
+                .editSpec()
+                    .editContainer(0)
+                        .withResources(new ResourceRequirementsBuilder()
+                                .withRequests(Map.of("cpu", new Quantity("1"), "memory", new Quantity("1024Mi")))
+                                .withLimits(Map.of("cpu", new Quantity("3"), "memory", new Quantity("4096Mi")))
+                                .build())
+                    .endContainer()
+                .endSpec()
+                .build();
+        assertThat(PodRevision.getRevision(Reconciliation.DUMMY_RECONCILIATION, pod2), is(basicPodRevision));
+
+        // Different pods have different revisions
+        pod2 = new PodBuilder(POD)
                 .editSpec()
                     .withTerminationGracePeriodSeconds(30L)
                 .endSpec()
@@ -64,9 +83,39 @@ public class PodRevisionTest {
     }
 
     @Test
+    public void testResourceRevisions() {
+        String basicPodRevision = PodRevision.getResourceRevision(Reconciliation.DUMMY_RECONCILIATION, POD);
+
+        // The same pod has always the same revision
+        assertThat(PodRevision.getResourceRevision(Reconciliation.DUMMY_RECONCILIATION, POD), is(basicPodRevision));
+        assertThat(PodRevision.getResourceRevision(Reconciliation.DUMMY_RECONCILIATION, new PodBuilder(POD).build()), is(basicPodRevision));
+
+        // When something else changes and not resources, the resource revision should not change
+        Pod pod2 = new PodBuilder(POD)
+                .editSpec()
+                    .withTerminationGracePeriodSeconds(30L)
+                .endSpec()
+                .build();
+        assertThat(PodRevision.getResourceRevision(Reconciliation.DUMMY_RECONCILIATION, pod2), is(basicPodRevision));
+
+        // Revision changes when resources change
+        pod2 = new PodBuilder(POD)
+                .editSpec()
+                    .editContainer(0)
+                        .withResources(new ResourceRequirementsBuilder()
+                                .withRequests(Map.of("cpu", new Quantity("1"), "memory", new Quantity("1024Mi")))
+                                .withLimits(Map.of("cpu", new Quantity("3"), "memory", new Quantity("4096Mi")))
+                                .build())
+                    .endContainer()
+                .endSpec()
+                .build();
+        assertThat(PodRevision.getResourceRevision(Reconciliation.DUMMY_RECONCILIATION, pod2), is(not(basicPodRevision)));
+    }
+
+    @Test
     public void testHasChangedWithPods()    {
         // Two pods without the revision annotation
-        assertThat(PodRevision.hasChanged(POD, new PodBuilder(POD).build()), is(false));
+        assertThat(PodRevision.hasChanged(POD, new PodBuilder(POD).build(), PodRevision.STRIMZI_REVISION_ANNOTATION), is(false));
 
         // Pods with the annotation
         Pod pod1 = new PodBuilder(POD)
@@ -80,15 +129,15 @@ public class PodRevisionTest {
                 .endMetadata()
                 .build();
 
-        assertThat(PodRevision.hasChanged(pod1, new PodBuilder(pod1).build()), is(false));
-        assertThat(PodRevision.hasChanged(pod1, pod2), is(true));
-        assertThat(PodRevision.hasChanged(POD, pod2), is(true));
+        assertThat(PodRevision.hasChanged(pod1, new PodBuilder(pod1).build(), PodRevision.STRIMZI_REVISION_ANNOTATION), is(false));
+        assertThat(PodRevision.hasChanged(pod1, pod2, PodRevision.STRIMZI_REVISION_ANNOTATION), is(true));
+        assertThat(PodRevision.hasChanged(POD, pod2, PodRevision.STRIMZI_REVISION_ANNOTATION), is(true));
     }
 
     @Test
     public void testHasChangedWithPodAndPodSet()    {
         // Two pods without the revision annotation
-        assertThat(PodRevision.hasChanged(POD, podSet(POD)), is(false));
+        assertThat(PodRevision.hasChanged(POD, podSet(POD), PodRevision.STRIMZI_REVISION_ANNOTATION), is(false));
 
         // Pods with the annotation
         Pod pod1 = new PodBuilder(POD)
@@ -108,10 +157,10 @@ public class PodRevisionTest {
                 .endMetadata()
                 .build();
 
-        assertThat(PodRevision.hasChanged(pod1, podSet(pod1)), is(false));
-        assertThat(PodRevision.hasChanged(pod1, podSet(pod1, pod3)), is(false));
-        assertThat(PodRevision.hasChanged(pod1, podSet(pod2)), is(true));
-        assertThat(PodRevision.hasChanged(POD, podSet(pod2)), is(true));
+        assertThat(PodRevision.hasChanged(pod1, podSet(pod1), PodRevision.STRIMZI_REVISION_ANNOTATION), is(false));
+        assertThat(PodRevision.hasChanged(pod1, podSet(pod1, pod3), PodRevision.STRIMZI_REVISION_ANNOTATION), is(false));
+        assertThat(PodRevision.hasChanged(pod1, podSet(pod2), PodRevision.STRIMZI_REVISION_ANNOTATION), is(true));
+        assertThat(PodRevision.hasChanged(POD, podSet(pod2), PodRevision.STRIMZI_REVISION_ANNOTATION), is(true));
     }
 
     /**