+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/operator/assembly/StrimziPodSetControllerMockTest.java
@@ -10,6 +10,7 @@ import io.fabric8.kubernetes.api.model.LabelSelector;
 import io.fabric8.kubernetes.api.model.OwnerReference;
 import io.fabric8.kubernetes.api.model.Pod;
 import io.fabric8.kubernetes.api.model.PodBuilder;
+import io.fabric8.kubernetes.api.model.Quantity;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.dsl.MixedOperation;
 import io.fabric8.kubernetes.client.dsl.Resource;
@@ -36,6 +37,7 @@ import io.strimzi.operator.cluster.model.PodSetUtils;
 import io.strimzi.operator.cluster.operator.resource.kubernetes.CrdOperator;
 import io.strimzi.operator.cluster.operator.resource.kubernetes.PodOperator;
 import io.strimzi.operator.cluster.operator.resource.kubernetes.StrimziPodSetOperator;
+import io.strimzi.operator.common.Annotations;
 import io.strimzi.operator.common.MetricsProvider;
 import io.strimzi.operator.common.Reconciliation;
 import io.strimzi.operator.common.metrics.MetricsHolder;
@@ -223,6 +225,10 @@ public class StrimziPodSetControllerMockTest {
                                 .withImage("quay.io/scholzj/busybox:latest") // Quay.io is used to avoid Docker Hub limits
                                 .withCommand("sleep", "3600")
                                 .withImagePullPolicy("IfNotPresent")
+                                .withNewResources()
+                                    .withRequests(Map.of("cpu", new Quantity("50m"), "memory", new Quantity("50Mi")))
+                                    .withLimits(Map.of("cpu", new Quantity("70m"), "memory", new Quantity("70Mi")))
+                                .endResources()
                                 .build())
                         .withRestartPolicy("Always")
                         .withTerminationGracePeriodSeconds(0L)
@@ -230,6 +236,7 @@ public class StrimziPodSetControllerMockTest {
                     .build();
 
         pod.getMetadata().getAnnotations().put(PodRevision.STRIMZI_REVISION_ANNOTATION, PodRevision.getRevision(Reconciliation.DUMMY_RECONCILIATION, pod));
+        pod.getMetadata().getAnnotations().put(PodRevision.STRIMZI_RESOURCE_REVISION_ANNOTATION, PodRevision.getResourceRevision(Reconciliation.DUMMY_RECONCILIATION, pod));
 
         return pod;
     }
@@ -794,6 +801,93 @@ public class StrimziPodSetControllerMockTest {
 
             // Check the Pod is not failed anymore
             assertThat(recreatedPod.get().getStatus().getPhase(), is(not("Failed")));
+
+            context.completeNow();
+        } finally {
+            podSetOp().inNamespace(namespace).withName(podSetName).delete();
+        }
+    }
+
+    /**
+     * Tests in-place updates to Kubernetes resources
+     *
+     * @param context   Test context
+     */
+    @Test
+    public void testInPlacePodResizing(VertxTestContext context) {
+        String podSetName = "pod-resizing";
+        String podName = podSetName + "-0";
+
+        try {
+            Pod originalPod = pod(namespace, podName, KAFKA_NAME, podSetName, "Kafka");
+            StrimziPodSet originalPodSet = new StrimziPodSetBuilder(podSet(namespace, podSetName, KAFKA_NAME, "Kafka", originalPod))
+                    .editMetadata()
+                        .addToAnnotations(Annotations.ANNO_STRIMZI_IO_IN_PLACE_RESIZING, "true")
+                    .endMetadata()
+                    .build();
+            podSetOp().inNamespace(namespace).resource(originalPodSet).create();
+
+            // Wait until the pod is ready
+            TestUtils.waitFor(
+                    "Wait for Pod to be ready",
+                    100,
+                    10_000,
+                    () -> client.pods().inNamespace(namespace).withName(podName).isReady(),
+                    () -> context.failNow("Test timed out waiting for pod readiness!"));
+
+            // Check status of the PodSet
+            TestUtils.waitFor(
+                    "Wait for StrimziPodSetStatus",
+                    100,
+                    10_000,
+                    () -> {
+                        StrimziPodSet podSet = podSetOp().inNamespace(namespace).withName(podSetName).get();
+                        return podSet.getStatus().getCurrentPods() == 1
+                                && podSet.getStatus().getReadyPods() == 1
+                                && podSet.getStatus().getPods() == 1;
+                    },
+                    () -> context.failNow("Pod stats do not match"));
+
+            // Update the pod with a new resources and resource revision
+            Pod updatedPod = new PodBuilder(pod(namespace, podName, KAFKA_NAME, podSetName, "Kafka"))
+                    .editMetadata()
+                        .removeFromAnnotations(PodRevision.STRIMZI_RESOURCE_REVISION_ANNOTATION)
+                        .addToAnnotations(PodRevision.STRIMZI_RESOURCE_REVISION_ANNOTATION, "new-revision")
+                    .endMetadata()
+                    .editSpec()
+                        .editFirstContainer()
+                            .withNewResources()
+                                .withRequests(Map.of("cpu", new Quantity("50m"), "memory", new Quantity("50Mi")))
+                                .withLimits(Map.of("cpu", new Quantity("75m"), "memory", new Quantity("75Mi")))
+                            .endResources()
+                        .endContainer()
+                    .endSpec()
+                    .build();
+            StrimziPodSet updatedPodSet = new StrimziPodSetBuilder(podSet(namespace, podSetName, KAFKA_NAME, "Kafka", updatedPod))
+                    .editMetadata()
+                        .addToAnnotations(Annotations.ANNO_STRIMZI_IO_IN_PLACE_RESIZING, "true")
+                    .endMetadata()
+                    .build();
+            podSetOp().inNamespace(namespace).resource(updatedPodSet).update();
+
+            // Wait or the Pod to be resized
+            TestUtils.waitFor(
+                    "Wait for Pod update",
+                    100,
+                    10_000,
+                    () -> {
+                        Pod pod = client.pods().inNamespace(namespace).withName(podName).get();
+                        return "new-revision".equals(pod.getMetadata().getAnnotations().get(PodRevision.STRIMZI_RESOURCE_REVISION_ANNOTATION));
+                    },
+                    () -> context.failNow("Pod revisions do not match"));
+
+            // Note: MockKube would not really resize the Pod. So we do not check the container statuses to make sure it was really resized.
+
+            // Check the pod was resized
+            Pod actualPod = client.pods().inNamespace(namespace).withName(podName).get();
+            assertThat(actualPod.getMetadata().getAnnotations().get(PodRevision.STRIMZI_RESOURCE_REVISION_ANNOTATION), is("new-revision"));
+            assertThat(actualPod.getSpec().getContainers().get(0).getResources().getRequests(), is(Map.of("cpu", new Quantity("50m"), "memory", new Quantity("50Mi"))));
+            assertThat(actualPod.getSpec().getContainers().get(0).getResources().getLimits(), is(Map.of("cpu", new Quantity("75m"), "memory", new Quantity("75Mi"))));
 
             context.completeNow();
         } finally {